Split code in smaller files
