+++ package.json
@@ -35,6 +35,7 @@
     "ts-node": "^9.1.1",
     "typescript": "^3.9.7"
   },
+  "dependencies": {},
   "repository": {
     "type": "git",
     "url": "git+https://github.com/gliese1337/fast-myers-diff.git"
